wip: try to sort gates
Encoding/Decoding fails because we recover a permutations of the
original bytes. This is probably because there is no guarantee that the
operations in a QuantumCircuit or DAGCircuit are always returned in the
same order. So we need to figure out a way to sort them without using
extry bytes for that.

Idea: Probably we can define an order base on metadata that persists
across different stages of the transpilation, e.g., the physical qubits
where gates are applied and the order within that qubit.