Use topological_nodes() instead of op_nodes()
The decoder was computing the right bytes but some in the wrong order.
It seems like DAGCircuit.op_nodes() is not deterministic in the sense
that the order of the operations might change from the time the data is
encoded until it arrives to the decoder. Using
DAGCircuit.topological_nodes() fixes this issue making the attack
possible.