Fix attack for recent versions of qiskis
Qiskit 1.2.0 moved many of the traspiler operations to Rust. This led to
the side effect that direct modifications of the DAGOpNode.op operations
are not saved because that object is now a copy.

This means that, in our attack, the modified rotations numbers encoding
the data were lost when the scheduling pass was finished.

The correct way to make this work is to use the
DAGCircuit.substitute_node() method to replace the operations in a safe
way.

Another issue had to do with the order returned by the decoder. This is
probably due to a bug in Qiskit, in either dag_to_circuit() or
circuit_to_dag() converters. The workaround is to convert the DAGCircuit
to a QuantumCircuit and then back to a DAGCircuit, so that the order of
the RZ gates is exactly the same when we encode and when we decode the
data.

See also: https://docs.quantum.ibm.com/api/qiskit/release-notes/1.2